+++ classes/Generator/Task/Generate/Trait.php
@@ -24,13 +24,6 @@ class Generator_Task_Generate_Trait extends Task_Generate_Class
 	);
 
 	/**
-	 * @var  array  Arguments mapped to options
-	 */
-	protected $_arguments = array(
-		1 => 'name',
-	);
-
-	/**
 	 * Validates the task options.
 	 *
 	 * @param   Validation  $validation  The validation object to add rules to